Remove unused code and redundant comment, move entity class
